Split JavaScript files to seperate cocoon required jquery into seperate file
