+++ app/javascript/packs/application.js
@@ -1,8 +1,3 @@
 //=require bootstrap
-import {
-  createNewCargoItemForm,
-  editCargoItemForm
-} from "../views/fixture_cargo_add";
 
-createNewCargoItemForm();
-editCargoItemForm();
+import { editCargoItemForm } from "../views/fixture_cargo_edit";
